@@ -1,14 +1,13 @@
 ﻿using MediatR;
 
-namespace Eshop.Application.Configuration.Commands
-{
-    public interface ICommand : IRequest
-    {
-        Guid Id { get; }
-    }
+namespace Eshop.Application.Configuration.Commands;
 
-    public interface ICommand<out TResult> : IRequest<TResult>
-    {
-        Guid Id { get; }
-    }
+public interface ICommand : IRequest
+{
+    Guid Id { get; }
+}
+
+public interface ICommand<out TResult> : IRequest<TResult>
+{
+    Guid Id { get; }
 }
