@@ -1,16 +1,15 @@
 ﻿using MediatR;
 
-namespace Eshop.Application.Configuration.Commands
+namespace Eshop.Application.Configuration.Commands;
+
+public interface ICommandHandler<in TCommand> :
+    IRequestHandler<TCommand> where TCommand : ICommand
 {
-    public interface ICommandHandler<in TCommand> :
-        IRequestHandler<TCommand> where TCommand : ICommand
-    {
 
-    }
+}
 
-    public interface ICommandHandler<in TCommand, TResult> :
-        IRequestHandler<TCommand, TResult> where TCommand : ICommand<TResult>
-    {
+public interface ICommandHandler<in TCommand, TResult> :
+    IRequestHandler<TCommand, TResult> where TCommand : ICommand<TResult>
+{
 
-    }
 }
