@@ -1,9 +1,8 @@
 ﻿using MediatR;
 
-namespace Eshop.Application.Configuration.Queries
-{
-    public interface IQuery<out TResult> : IRequest<TResult>
-    {
+namespace Eshop.Application.Configuration.Queries;
 
-    }
+public interface IQuery<out TResult> : IRequest<TResult>
+{
+
 }
