@@ -1,10 +1,9 @@
 ﻿using MediatR;
 
-namespace Eshop.Application.Configuration.Queries
-{
-    public interface IQueryHandler<in TQuery, TResult> :
-        IRequestHandler<TQuery, TResult> where TQuery : IQuery<TResult>
-    {
+namespace Eshop.Application.Configuration.Queries;
 
-    }
+public interface IQueryHandler<in TQuery, TResult> :
+    IRequestHandler<TQuery, TResult> where TQuery : IQuery<TResult>
+{
+
 }
