@@ -1,14 +1,11 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 
-namespace Eshop.Application
-{
-    public static class Registry
-    {
-        public static void RegistryApplication(this IServiceCollection services)
-        {
-            services.AddAutoMapper(typeof(Registry));
+namespace Eshop.Application;
 
-            services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(Registry).Assembly));
-        }
+public static class Registry
+{
+    public static void RegistryApplication(this IServiceCollection services)
+    {
+        services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(Registry).Assembly));
     }
 }
