@@ -1,17 +1,10 @@
 ﻿using Eshop.Domain.SeedWork;
 
-namespace Eshop.Domain.Orders.Events
+namespace Eshop.Domain.Orders.Events;
+
+public class OrderAddedEvent(Guid orderId, Guid customerId) : DomainEventBase
 {
-    public class OrderAddedEvent : DomainEventBase
-    {
-        public Guid OrderId { get; }
+    public Guid OrderId { get; } = orderId;
 
-        public Guid CustomerId { get; }
-
-        public OrderAddedEvent(Guid orderId, Guid customerId)
-        {
-            OrderId = orderId;
-            CustomerId = customerId;
-        }
-    }
+    public Guid CustomerId { get; } = customerId;
 }
