@@ -1,9 +1,8 @@
-﻿namespace Eshop.Domain.Orders
-{
-    public interface IOrderRepository
-    {
-        Task<Order> GetByIdAsync(Guid id);
+﻿namespace Eshop.Domain.Orders;
 
-        void Add(Order order);
-    }
+public interface IOrderRepository
+{
+    Task<Order> GetByIdAsync(Guid id);
+
+    void Add(Order order);
 }
