@@ -1,18 +1,10 @@
 ﻿using Eshop.Domain.SeedWork;
 
-namespace Eshop.Domain.Orders.Rules
+namespace Eshop.Domain.Orders.Rules;
+
+public class OrderMustHaveAtLeastOneProductRule(IReadOnlyCollection<OrderProduct> orderProducts) : IBusinessRule
 {
-    public class OrderMustHaveAtLeastOneProductRule : IBusinessRule
-    {
-        private readonly List<OrderProduct> _orderProducts;
+    public bool IsBroken() => orderProducts.Count == 0;
 
-        public OrderMustHaveAtLeastOneProductRule(List<OrderProduct> orderProducts)
-        {
-            _orderProducts = orderProducts;
-        }
-
-        public bool IsBroken() => !_orderProducts.Any();
-
-        public string Message => "Order must have at least one product";
-    }
+    public string Message => "Order must have at least one product";
 }
