@@ -1,17 +1,10 @@
 ﻿using Eshop.Domain.SeedWork;
 
-namespace Eshop.Domain.Products
+namespace Eshop.Domain.Products;
+
+public class ProductPriceData(Guid productId, decimal unitPrice) : ValueObject
 {
-    public class ProductPriceData : ValueObject
-    {
-        public ProductPriceData(Guid productId, decimal unitPrice)
-        {
-            ProductId = productId;
-            UnitPrice = unitPrice;
-        }
+    public Guid ProductId { get; } = productId;
 
-        public Guid ProductId { get; }
-
-        public decimal UnitPrice { get; set; }
-    }
+    public decimal UnitPrice { get; set; } = unitPrice;
 }
