@@ -1,41 +1,46 @@
-﻿namespace Eshop.Domain.SeedWork
+﻿using MongoDB.Bson;
+using MongoDB.Bson.Serialization.Attributes;
+
+namespace Eshop.Domain.SeedWork;
+
+/// <summary>
+/// Base class for entities.
+/// </summary>
+public abstract class Entity(Guid id)
 {
+    [BsonId]
+    [BsonRepresentation(BsonType.String)]
+    public Guid Id { get; private set; } = id;
+
+    private readonly List<IDomainEvent> _domainEvents = [];
+
     /// <summary>
-    /// Base class for entities.
+    /// Domain events occurred.
     /// </summary>
-    public abstract class Entity
+    public IReadOnlyCollection<IDomainEvent> DomainEvents => _domainEvents?.AsReadOnly() ?? new List<IDomainEvent>().AsReadOnly();
+
+    /// <summary>
+    /// Add domain event.
+    /// </summary>
+    /// <param name="domainEvent"></param>
+    protected void AddDomainEvent(IDomainEvent domainEvent)
     {
-        private List<IDomainEvent>? _domainEvents;
+        _domainEvents.Add(domainEvent);
+    }
 
-        /// <summary>
-        /// Domain events occurred.
-        /// </summary>
-        public IReadOnlyCollection<IDomainEvent> DomainEvents => _domainEvents?.AsReadOnly() ?? new List<IDomainEvent>().AsReadOnly();
+    /// <summary>
+    /// Clear domain events.
+    /// </summary>
+    public void ClearDomainEvents()
+    {
+        _domainEvents?.Clear();
+    }
 
-        /// <summary>
-        /// Add domain event.
-        /// </summary>
-        /// <param name="domainEvent"></param>
-        protected void AddDomainEvent(IDomainEvent domainEvent)
+    protected static void CheckRule(IBusinessRule rule)
+    {
+        if (rule.IsBroken())
         {
-            _domainEvents = _domainEvents ?? new List<IDomainEvent>();
-            _domainEvents.Add(domainEvent);
-        }
-
-        /// <summary>
-        /// Clear domain events.
-        /// </summary>
-        public void ClearDomainEvents()
-        {
-            _domainEvents?.Clear();
-        }
-
-        protected static void CheckRule(IBusinessRule rule)
-        {
-            if (rule.IsBroken())
-            {
-                throw new BusinessRuleValidationException(rule);
-            }
+            throw new BusinessRuleValidationException(rule);
         }
     }
 }
