@@ -1,7 +1,6 @@
-﻿namespace Eshop.Domain.SeedWork
+﻿namespace Eshop.Domain.SeedWork;
+
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Field)]
+public sealed class IgnoreMemberAttribute : Attribute
 {
-    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field)]
-    public sealed class IgnoreMemberAttribute : Attribute
-    {
-    }
 }
