@@ -1,45 +1,39 @@
-﻿namespace Eshop.Domain.SeedWork
+﻿namespace Eshop.Domain.SeedWork;
+
+public abstract class TypedIdValueBase(Guid value) : IEquatable<TypedIdValueBase>
 {
-    public abstract class TypedIdValueBase : IEquatable<TypedIdValueBase>
+    public Guid Value { get; } = value;
+
+    public override bool Equals(object obj)
     {
-        public Guid Value { get; }
+        if (ReferenceEquals(null, obj)) return false;
+        return obj is TypedIdValueBase other && Equals(other);
+    }
 
-        protected TypedIdValueBase(Guid value)
-        {
-            Value = value;
-        }
+    public override int GetHashCode()
+    {
+        return Value.GetHashCode();
+    }
 
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            return obj is TypedIdValueBase other && Equals(other);
-        }
+    public bool Equals(TypedIdValueBase other)
+    {
+        return Value == other.Value;
+    }
 
-        public override int GetHashCode()
+    public static bool operator ==(TypedIdValueBase obj1, TypedIdValueBase obj2)
+    {
+        if (object.Equals(obj1, null))
         {
-            return Value.GetHashCode();
-        }
-
-        public bool Equals(TypedIdValueBase other)
-        {
-            return Value == other.Value;
-        }
-
-        public static bool operator ==(TypedIdValueBase obj1, TypedIdValueBase obj2)
-        {
-            if (object.Equals(obj1, null))
+            if (object.Equals(obj2, null))
             {
-                if (object.Equals(obj2, null))
-                {
-                    return true;
-                }
-                return false;
+                return true;
             }
-            return obj1.Equals(obj2);
+            return false;
         }
-        public static bool operator !=(TypedIdValueBase x, TypedIdValueBase y) 
-        {
-            return !(x == y);
-        }
+        return obj1.Equals(obj2);
+    }
+    public static bool operator !=(TypedIdValueBase x, TypedIdValueBase y) 
+    {
+        return !(x == y);
     }
 }
