@@ -1,115 +1,114 @@
 ﻿using System.Reflection;
 
-namespace Eshop.Domain.SeedWork
+namespace Eshop.Domain.SeedWork;
+
+public abstract class ValueObject : IEquatable<ValueObject>
 {
-    public abstract class ValueObject : IEquatable<ValueObject>
+    private List<PropertyInfo> _properties;
+    private List<FieldInfo> _fields;
+
+    public static bool operator ==(ValueObject obj1, ValueObject obj2)
     {
-        private List<PropertyInfo> _properties;
-        private List<FieldInfo> _fields;
-
-        public static bool operator ==(ValueObject obj1, ValueObject obj2)
+        if (object.Equals(obj1, null))
         {
-            if (object.Equals(obj1, null))
+            if (object.Equals(obj2, null))
             {
-                if (object.Equals(obj2, null))
-                {
-                    return true;
-                }
-                return false;
+                return true;
             }
-            return obj1.Equals(obj2);
+            return false;
+        }
+        return obj1.Equals(obj2);
+    }
+
+    public static bool operator !=(ValueObject obj1, ValueObject obj2)
+    {
+        return !(obj1 == obj2);
+    }
+
+    public bool Equals(ValueObject obj)
+    {
+        return Equals(obj as object);
+    }
+
+    public override bool Equals(object obj)
+    {
+        if (obj == null || GetType() != obj.GetType()) return false;
+
+        return GetProperties().All(p => PropertiesAreEqual(obj, p))
+               && GetFields().All(f => FieldsAreEqual(obj, f));
+    }
+
+    private bool PropertiesAreEqual(object obj, PropertyInfo p)
+    {
+        return object.Equals(p.GetValue(this, null), p.GetValue(obj, null));
+    }
+
+    private bool FieldsAreEqual(object obj, FieldInfo f)
+    {
+        return object.Equals(f.GetValue(this), f.GetValue(obj));
+    }
+
+    private IEnumerable<PropertyInfo> GetProperties()
+    {
+        if (_properties == null)
+        {
+            _properties = GetType()
+                .GetProperties(BindingFlags.Instance | BindingFlags.Public)
+                .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
+                .ToList();
+
+            // Not available in Core
+            // !Attribute.IsDefined(p, typeof(IgnoreMemberAttribute))).ToList();
         }
 
-        public static bool operator !=(ValueObject obj1, ValueObject obj2)
+        return _properties;
+    }
+
+    private IEnumerable<FieldInfo> GetFields()
+    {
+        if (_fields == null)
         {
-            return !(obj1 == obj2);
+            _fields = GetType().GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)
+                .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
+                .ToList();
         }
 
-        public bool Equals(ValueObject obj)
-        {
-            return Equals(obj as object);
-        }
+        return _fields;
+    }
 
-        public override bool Equals(object obj)
+    public override int GetHashCode()
+    {
+        unchecked   //allow overflow
         {
-            if (obj == null || GetType() != obj.GetType()) return false;
-
-            return GetProperties().All(p => PropertiesAreEqual(obj, p))
-                && GetFields().All(f => FieldsAreEqual(obj, f));
-        }
-
-        private bool PropertiesAreEqual(object obj, PropertyInfo p)
-        {
-            return object.Equals(p.GetValue(this, null), p.GetValue(obj, null));
-        }
-
-        private bool FieldsAreEqual(object obj, FieldInfo f)
-        {
-            return object.Equals(f.GetValue(this), f.GetValue(obj));
-        }
-
-        private IEnumerable<PropertyInfo> GetProperties()
-        {
-            if (_properties == null)
+            int hash = 17;
+            foreach (var prop in GetProperties())
             {
-                _properties = GetType()
-                    .GetProperties(BindingFlags.Instance | BindingFlags.Public)
-                    .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
-                    .ToList();
-
-                // Not available in Core
-                // !Attribute.IsDefined(p, typeof(IgnoreMemberAttribute))).ToList();
+                var value = prop.GetValue(this, null);
+                hash = HashValue(hash, value);
             }
 
-            return _properties;
-        }
-
-        private IEnumerable<FieldInfo> GetFields()
-        {
-            if (_fields == null)
+            foreach (var field in GetFields())
             {
-                _fields = GetType().GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)
-                    .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
-                    .ToList();
+                var value = field.GetValue(this);
+                hash = HashValue(hash, value);
             }
 
-            return _fields;
+            return hash;
         }
+    }
 
-        public override int GetHashCode()
+    private static int HashValue(int seed, object value)
+    {
+        var currentHash = value?.GetHashCode() ?? 0;
+
+        return seed * 23 + currentHash;
+    }
+
+    protected static void CheckRule(IBusinessRule rule)
+    {
+        if (rule.IsBroken())
         {
-            unchecked   //allow overflow
-            {
-                int hash = 17;
-                foreach (var prop in GetProperties())
-                {
-                    var value = prop.GetValue(this, null);
-                    hash = HashValue(hash, value);
-                }
-
-                foreach (var field in GetFields())
-                {
-                    var value = field.GetValue(this);
-                    hash = HashValue(hash, value);
-                }
-
-                return hash;
-            }
-        }
-
-        private static int HashValue(int seed, object value)
-        {
-            var currentHash = value?.GetHashCode() ?? 0;
-
-            return seed * 23 + currentHash;
-        }
-
-        protected static void CheckRule(IBusinessRule rule)
-        {
-            if (rule.IsBroken())
-            {
-                throw new BusinessRuleValidationException(rule);
-            }
+            throw new BusinessRuleValidationException(rule);
         }
     }
 }
