@@ -1,8 +1,7 @@
-﻿namespace Eshop.Infrastructure.Database
+﻿namespace Eshop.Infrastructure.Database;
+
+internal class MongoDbSettings
 {
-    internal class MongoDbSettings
-    {
-        public required string ConnectionString { get; set; }
-        public required string DatabaseName { get; set; }
-    }
+    public required string ConnectionString { get; init; }
+    public required string DatabaseName { get; init; }
 }
