@@ -1,4 +1,5 @@
-﻿using Eshop.Domain.Orders;
+﻿using Eshop.Application.Orders;
+using Eshop.Domain.Orders;
 using Eshop.Domain.SeedWork;
 using Eshop.Infrastructure.Database;
 using Eshop.Infrastructure.Repositories;
@@ -6,33 +7,32 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using MongoDB.Driver;
 
-namespace Eshop.Infrastructure
+namespace Eshop.Infrastructure;
+
+public static class Registry
 {
-    public static class Registry
+    public static void RegistryInfrastructure(this IServiceCollection services, IConfiguration configuration)
     {
-        public static void RegistryInfrastructure(this IServiceCollection services, IConfiguration configuration)
+        services.AddScoped<IOrderRepository, OrderRepository>();
+        services.AddScoped<IProductPriceDataApi, ProductPriceDataApi>();
+        services.AddScoped<IUnitOfWork, UnitOfWork>();
+        services.AddScoped<IDomainEventsDispatcher, DomainEventsDispatcher>();
+        services.AddScoped<IEntityTracker, EntityTracker>();
+
+        var mongoDbSettings = configuration.GetSection("MongoDB").Get<MongoDbSettings>() ?? throw new InvalidOperationException("MongoDB settings are not configured properly.");
+            
+        if (string.IsNullOrEmpty(mongoDbSettings.ConnectionString))
+            throw new InvalidOperationException("MongoDB connection string is not configured.");
+            
+        services.AddSingleton<IMongoClient>(_ => new MongoClient(mongoDbSettings.ConnectionString));
+
+        services.AddSingleton(provider =>
         {
-            services.AddScoped<IOrderRepository, OrderRepository>();
-
-            services.AddScoped<IProductPriceDataApi, ProductPriceDataApi>();
-
-            services.AddScoped<IUnitOfWork, UnitOfWork>();
-
-            services.AddScoped<IEntityTracker, EntityTracker>();
-
-            services.AddScoped<IDomainEventsDispatcher, DomainEventsDispatcher>();
-
-            MongoDbSettings mongoDbSettings = configuration.GetSection("MongoDB").Get<MongoDbSettings>() ?? throw new InvalidOperationException();
-
-            services.AddSingleton<IMongoClient>(ServiceProvider =>
-            {
-                return new MongoClient(mongoDbSettings.ConnectionString);
-            });
-
-            services.AddSingleton(provider =>
-            {
-                return new OrdersContext(mongoDbSettings);
-            });
-        }
+            var client = provider.GetRequiredService<IMongoClient>();
+            if (string.IsNullOrEmpty(mongoDbSettings.DatabaseName))
+                throw new InvalidOperationException("MongoDB database name is not configured.");
+            var database = client.GetDatabase(mongoDbSettings.DatabaseName);
+            return new OrdersContext(database);
+        });
     }
 }
