@@ -1,9 +1,9 @@
-﻿public class OrderNotExistsException : Exception 
-{ 
-    public Guid Id { get; }
+﻿namespace Eshop.Infrastructure.Exceptions;
 
-    public OrderNotExistsException(Guid id)
-    {
-        Id = id;
-    }
+/// <summary>
+/// Exception thrown when an order with the specified ID does not exist.
+/// </summary>
+public class OrderNotExistsException(Guid id) : Exception($"Order with ID '{id}' does not exist.")
+{
+    public Guid Id { get; } = id;
 }
